#includeできないor問題のあるものに関してissueに一括記載 定数用インタプリタvalの型変更、cast演算子実装 配列定義の添字に直値以外の定数式を使用可能に #if がTrueの場合に#else #elifがスキップされないバグ修正

+++ src/9cc.h
@@ -348,7 +348,7 @@ void gen(Node *node);
 
 // type.c
 void init_sizes();
-int val(Node *node);  // 定数計算
+long long val(Node *node); // 定数計算
 int size(Type *typ);
 bool can_deref(Type *typ);
 bool eqtype(Type *typ1, Type *typ2);
